+++ welcome_message.rb
@@ -1,0 +1,11 @@
+def welcome_message
+    puts 'wELCOME TO THE SCHOOL LIBRARY APP!'
+    puts 'Please choose an option by enterin a number:'
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals for a given person id'
+    puts '7 - Exit'
+  end
